+++ src/Animals/RedFish.java
@@ -8,7 +8,7 @@ import Species.Fish;
  */
 public class RedFish extends Fish implements ISwim {
 
-    public RedFish(String name, boolean sex, int weight, int height, int age) {
+    public RedFish(String name, boolean sex, long weight, int height, int age) {
         super(name, sex, weight, height, age);
     } // Constructor
